Split funder into lib and bin
