+++ utils/funder/Cargo.toml
@@ -3,7 +3,13 @@ name = "funder"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "funder_lib"
+path = "src/lib.rs"
+
+[[bin]]
+name = "funder"
+path = "src/main.rs"
 
 [dependencies]
 utils = { path = ".." }
